moved files and stories to FormPage
